# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LRMC — Legacy Rental Management Consortium API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
            default: /
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL with a post-logout redirect back to the validated `returnTo` path.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  # ---------------- Users ----------------
  /users/me/profile:
    patch:
      tags: [membership]
      operationId: updateMyProfile
      summary: Update the current user's phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"

  # ---------------- Membership ----------------
...
  /assets:
    get:
      tags: [assets]
      operationId: listAssets
      summary: List property/vehicle/Airbnb/resort assets
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            type: string
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      tags: [assets]
      operationId: createAsset
      summary: List a new asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    patch:
      tags: [assets]
      operationId: updateAsset
      summary: Update an asset
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      tags: [assets]
      operationId: deleteAsset
      summary: Remove an asset
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ---------------- Land ----------------
...
  /ads:
    get:
      tags: [ads]
      operationId: listAds
      summary: List advertisements
      parameters:
        - name: advertiserId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      tags: [ads]
      operationId: createAd
      summary: Submit an advertisement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
  /ads/{id}:
    patch:
      tags: [ads]
      operationId: updateAd
      summary: Update an advertisement's status
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"

  # ---------------- Marketplace ----------------
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, role, phone]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
        phone:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    ProfileUpdate:
      type: object
      properties:
        phone:
          type: string
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ---- Membership ----
...
    Asset:
      type: object
      required: [id, ownerId, kind, title, description, location, priceDalasi, imageUrl, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        ownerId: { type: string }
        kind: { type: string }
        title: { type: string }
        description: { type: string }
        location: { type: string }
        priceDalasi: { type: integer }
        imageUrl: { type: [string, 'null'] }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AssetInput:
      type: object
      required: [ownerId, kind, title]
      properties:
        ownerId: { type: string }
        kind: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string }
        location: { type: string }
        priceDalasi: { type: integer }
        imageUrl: { type: string }
    AssetUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        location: { type: string }
        priceDalasi: { type: integer }
        imageUrl: { type: string }
        status: { type: string }

    # ---- Land ----
...
    Ad:
      type: object
      required: [id, advertiserId, title, description, placement, budgetUsd, status, createdAt]
      properties:
        id: { type: integer }
        advertiserId: { type: string }
        title: { type: string }
        description: { type: string }
        placement: { type: string }
        budgetUsd: { type: integer }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    AdInput:
      type: object
      required: [advertiserId, title]
      properties:
        advertiserId: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string }
        placement: { type: string }
        budgetUsd: { type: integer }
    AdUpdate:
      type: object
      properties:
        status: { type: string }
        budgetUsd: { type: integer }

    # ---- Marketplace ----
...